Fixed sleep times
