--- consensus/src/main/kotlin/maru/consensus/Web3jMetadataProvider.kt
@@ -1,43 +1,0 @@
-/*
-   Copyright 2025 Consensys Software Inc.
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
- */
-package maru.consensus
-
-import org.apache.tuweni.bytes.Bytes32
-import org.web3j.protocol.Web3j
-import org.web3j.protocol.core.DefaultBlockParameter
-import tech.pegasys.teku.infrastructure.async.SafeFuture
-
-class Web3jMetadataProvider(
-  private val web3jEthereumApiClient: Web3j,
-) {
-  fun getLatestBlockMetadata(): SafeFuture<BlockMetadata> =
-    SafeFuture
-      .of(
-        web3jEthereumApiClient
-          .ethGetBlockByNumber(DefaultBlockParameter.valueOf("latest"), false)
-          .sendAsync()
-          .minimalCompletionStage(),
-      ).thenApply {
-        val block = it.block
-        BlockMetadata(
-          block.number
-            .toLong()
-            .toULong(),
-          Bytes32.fromHexString(block.hash).toArray(),
-          block.timestamp.toLong(),
-        )
-      }
-}
