+++ consensus/src/main/kotlin/maru/consensus/qbft/Extensions.kt
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package maru.consensus
+package maru.consensus.qbft
 
 import maru.core.BeaconBlockHeader
 import maru.core.Validator
