+++ consensus/src/main/kotlin/maru/consensus/qbft/adapters/QbftFinalStateAdapter.kt
@@ -16,8 +16,8 @@
 package maru.consensus.qbft.adapters
 
 import java.time.Clock
+import maru.consensus.qbft.toAddress
 import maru.database.BeaconChain
-import org.apache.tuweni.bytes.Bytes
 import org.hyperledger.besu.consensus.common.bft.BftHelpers
 import org.hyperledger.besu.consensus.common.bft.BlockTimer
 import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier
@@ -51,9 +51,7 @@ class QbftFinalStateAdapter(
   override fun isLocalNodeValidator(): Boolean = getValidators().contains(localAddress)
 
   override fun getValidators(): Collection<Address> =
-    beaconChain.getLatestBeaconState().validators.map {
-      Address.wrap(Bytes.wrap(it.address))
-    }
+    beaconChain.getLatestBeaconState().validators.map { it.toAddress() }
 
   override fun getLocalAddress(): Address = localAddress
